# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Freedom.AI Operational Hub - API POC Viseu
  version: 0.1.0-poc
  description: >
    Contrato formal da API da POC. Escopo fechado para producao da POC:
    nao implementar alem do escopo contratado e nao deixar de implementar
    nenhuma funcionalidade do escopo.
servers:
  - url: http://localhost:3001
    description: Backend local
...
paths:
...
  /api/v1/prepostos:
    get:
      tags: [Prepostos]
      summary: Listar prepostos
      parameters:
        - in: query
          name: busca
          schema:
            type: string
        - in: query
          name: ativo
          schema:
            type: boolean
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Lista paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: "#/components/schemas/Preposto"
                  paginacao:
                    $ref: "#/components/schemas/Paginacao"
                required: [dados, paginacao]
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
    post:
      tags: [Prepostos]
      summary: Criar preposto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrepostoCreateRequest"
      responses:
        "201":
          description: Preposto criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preposto"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
  /api/v1/prepostos/{id}:
    patch:
      tags: [Prepostos]
      summary: Atualizar preposto
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrepostoUpdateRequest"
      responses:
        "200":
          description: Preposto atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preposto"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
    delete:
      tags: [Prepostos]
      summary: Remover preposto (soft delete)
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      responses:
        "200":
          description: Preposto removido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
                    format: uuid
                required: [message, id]
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/parceiros:
    get:
      tags: [Parceiros]
      summary: Listar parceiros
      parameters:
        - in: query
          name: busca
          schema:
            type: string
        - in: query
          name: ativo
          schema:
            type: boolean
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Lista paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: "#/components/schemas/Parceiro"
                  paginacao:
                    $ref: "#/components/schemas/Paginacao"
                required: [dados, paginacao]
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
    post:
      tags: [Parceiros]
      summary: Criar parceiro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParceiroCreateRequest"
      responses:
        "201":
          description: Parceiro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parceiro"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
  /api/v1/parceiros/{id}:
    patch:
      tags: [Parceiros]
      summary: Atualizar parceiro
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParceiroUpdateRequest"
      responses:
        "200":
          description: Parceiro atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parceiro"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
  /api/v1/parceiros/{id}/contatos:
    get:
      tags: [Parceiros]
      summary: Listar contatos do parceiro
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      responses:
        "200":
          description: Lista de contatos
          content:
            application/json:
              schema:
                type: object
                properties:
                  parceiro:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      nome:
                        type: string
                    required: [id, nome]
                  dados:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContatoParceiro"
                required: [parceiro, dados]
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
    post:
      tags: [Parceiros]
      summary: Criar contato do parceiro
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContatoCreateRequest"
      responses:
        "201":
          description: Contato criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContatoParceiro"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/importacoes:
    get:
      tags: [Importacoes]
      summary: Historico de importacoes
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Lista paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  dados:
                    type: array
                    items:
                      $ref: "#/components/schemas/Importacao"
                  paginacao:
                    $ref: "#/components/schemas/Paginacao"
                required: [dados, paginacao]
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /api/v1/importacoes/upload:
    post:
      tags: [Importacoes]
      summary: Upload de planilha
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "201":
          description: Upload realizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportacaoUploadResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /api/v1/importacoes/{id}/mapear:
    post:
      tags: [Importacoes]
      summary: Mapear colunas da planilha
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportacaoMapearRequest"
      responses:
        "200":
          description: Mapeamento salvo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportacaoMapearResponse"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/importacoes/{id}/preview:
    get:
      tags: [Importacoes]
      summary: Preview de validacao da importacao
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      responses:
        "200":
          description: Preview da importacao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportacaoPreviewResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/importacoes/{id}/confirmar:
    post:
      tags: [Importacoes]
      summary: Confirmar e efetivar importacao
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      responses:
        "200":
          description: Importacao concluida
          content:
            application/json:
              schema:
                type: object
                properties:
                  importacaoId:
                    type: string
                    format: uuid
                  status:
                    $ref: "#/components/schemas/StatusImportacao"
                  importadas:
                    type: integer
                  ignoradas:
                    type: integer
                  invalidas:
                    type: integer
                  ignoradasPorTrt:
                    type: integer
                  totalLinhas:
                    type: integer
                required:
                  - importacaoId
                  - status
                  - importadas
                  - ignoradas
                  - invalidas
                  - ignoradasPorTrt
                  - totalLinhas
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathIdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
  responses:
    BadRequestResponse:
      description: Requisicao invalida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationErrorResponse:
      description: Erro de validacao
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    UnauthorizedResponse:
      description: Nao autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFoundResponse:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ConflictResponse:
      description: Conflito de dados
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    StatusImportacao:
      type: string
      enum: [PROCESSANDO, CONCLUIDA, ERRO]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
        detalhes:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [error]
    Paginacao:
      type: object
      properties:
        total:
          type: integer
        pagina:
          type: integer
        limite:
          type: integer
        totalPaginas:
          type: integer
      required: [total, pagina, limite, totalPaginas]
...
    Preposto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        telefoneWhatsapp:
          type: string
        email:
          type: string
          nullable: true
        cpf:
          type: string
          nullable: true
        ativo:
          type: boolean
      required: [id, nome, telefoneWhatsapp, ativo]
    PrepostoCreateRequest:
      type: object
      properties:
        nome:
          type: string
        telefoneWhatsapp:
          type: string
        email:
          type: string
          format: email
        cpf:
          type: string
        ativo:
          type: boolean
      required: [nome, telefoneWhatsapp]
    PrepostoUpdateRequest:
      type: object
      properties:
        nome:
          type: string
        telefoneWhatsapp:
          type: string
        email:
          type: string
          format: email
        cpf:
          type: string
        ativo:
          type: boolean
      additionalProperties: false
    Parceiro:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        ativo:
          type: boolean
      required: [id, nome, ativo]
    ParceiroCreateRequest:
      type: object
      properties:
        nome:
          type: string
        ativo:
          type: boolean
      required: [nome]
    ParceiroUpdateRequest:
      type: object
      properties:
        nome:
          type: string
        ativo:
          type: boolean
      additionalProperties: false
    ContatoParceiro:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parceiroId:
          type: string
          format: uuid
        nome:
          type: string
        telefoneWhatsapp:
          type: string
        email:
          type: string
          nullable: true
        cargo:
          type: string
          nullable: true
        ordemEscalonamento:
          type: integer
      required: [id, parceiroId, nome, telefoneWhatsapp, ordemEscalonamento]
    ContatoCreateRequest:
      type: object
      properties:
        nome:
          type: string
        telefoneWhatsapp:
          type: string
        email:
          type: string
          format: email
        cargo:
          type: string
        ordemEscalonamento:
          type: integer
          minimum: 1
      required: [nome, telefoneWhatsapp]
...
    Importacao:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nomeArquivo:
          type: string
        totalRegistros:
          type: integer
        registrosImportados:
          type: integer
        registrosIgnorados:
          type: integer
        status:
          $ref: "#/components/schemas/StatusImportacao"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nomeArquivo
        - totalRegistros
        - registrosImportados
        - registrosIgnorados
        - status
        - createdAt
    ImportacaoUploadResponse:
      type: object
      properties:
        importacaoId:
          type: string
          format: uuid
        nomeArquivo:
          type: string
        totalRegistros:
          type: integer
        colunasDetectadas:
          type: array
          items:
            type: string
        previewLinhas:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - importacaoId
        - nomeArquivo
        - totalRegistros
        - colunasDetectadas
        - previewLinhas
    ImportacaoMapearRequest:
      type: object
      properties:
        mapeamentoColunas:
          type: object
          additionalProperties:
            type: string
          description: >
            Campos de dominio mapeados para colunas reais da planilha.
            Campos obrigatorios: numeroProcesso, data, trt, prepostoNome, parceiroNome.
      required: [mapeamentoColunas]
    ImportacaoMapearResponse:
      type: object
      properties:
        sucesso:
          type: boolean
        importacaoId:
          type: string
          format: uuid
        mapeamentoColunas:
          type: object
          additionalProperties:
            type: string
        status:
          $ref: "#/components/schemas/StatusImportacao"
        camposFaltantes:
          type: array
          items:
            type: string
      required: [sucesso]
    ImportacaoPreviewResponse:
      type: object
      properties:
        importacaoId:
          type: string
          format: uuid
        nomeArquivo:
          type: string
        totalLinhas:
          type: integer
        validasParaImportacao:
          type: integer
        ignoradasPorTrt:
          type: integer
        invalidas:
          type: integer
        erros:
          type: array
          items:
            type: object
            properties:
              linha:
                type: integer
              erros:
                type: array
                items:
                  type: string
            required: [linha, erros]
        amostraValidas:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - importacaoId
        - nomeArquivo
        - totalLinhas
        - validasParaImportacao
        - ignoradasPorTrt
        - invalidas
        - erros
        - amostraValidas
security:
  - bearerAuth: []